More work on the USB camera focus lock. This focus lock now tracks two spots at once and the error is the distance between them. The idea is that this will provide common mode rejection for things like pointing stability issues.
The focus lock spots are created by a IR diode laser. This laser (I think) hops between lasing at different wavelengths, which causes strange (and time varying) interference patterns between the spots. I found that I could resolve this issue by turning the laser on and off really rapidly so that it would (on the time scale of the camera) average across all the different interference patterns it could generate. This makes the spots smoother and easier to fit accurately. As a side effect you can now control the IR laser power from the computer by setting the duty cycle and the on/off digital line is no longer needed.