This is a fairly major change in the organization of the code. In an attempt to make it easier to control two cameras with one piece of software I have created a new "master" camera class (camera.genericCamera is generic base class) that interfaces between the main HAL class and the previous cameraControl, cameraDisplay and cameraParams classes. The purpose of this class, in so far as this is possible, is to move much of the stuff related to the control/display of data from a single camera out of the main HAL class so that additional cameras can be controlled without extensive modifications, tons of if statements and other ugliness in the HAL class. Following what is done for the illuminationControl, etc. it is now necessary to write a master camera class file for each setup (noneCamera, storm3Camera, etc.). For all of our current single objective setups all that should be necessary is for this class to be sub-class of camera.classicSingleCamera (see e.g. noneCamera.py for an example).
A frame of camera data is now an object with data that includes it's type (i.e. from camera1, camera2), its number (so that it is no longer necessary for all the different classes to keep their own record of this) and the data.

In addition I have attempted a further regularization of the names of functions.

These changes have only been (lightly) tested in the none emulation mode, so there are likely a number of bugs still in the code.