Change imagewriter classes so that they can handle saving data that is generated by more than one camera. Basically you will get one file per camera.
Change cameraDisplay class so that it handles all of the image display stuff itself, instead of relying on another class to feed it images & tell it when to update the display. This simplifies some of the camera stuff as the camera class now only has to call the newFrames method with the frames and doesn't have to deal with all the sync calculations, which camera to display, etc. 

Change Record button so that it disabled, but does not disappear in TCP mode. When it was hidden, other elements of the UI would shift around, which was a unexpected behavior.